+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@ A new header is inserted each time a *tag* is created.
 - Added ETC2 and BC compressed texture support to Metal backend.
 - Rendering a SAMPLER_EXTERNAL texture before setting an external image no longer results in GPU errors.
 - Fixed a normals issue when skinning without a normal map or anisotropy.
+- Fixed an issue where translucent views couldn't be used with post-processing.
 
 ## v1.4.2